chore(new button): factors out the newButton component PE-2237
